# Add centrality performance benchmarks