After 500 iterations for in-degree, and out-degree, and 10 iterations for closeness:
<table>
  <tr>
    <td></td><td><b>In-Degree</b></td><td><b>Out-Degree</b></td><td><b>Closeness</b></td>
  </tr>
  <tr>
    <td><b>Facebook</b></td><td>268 &mu;s</td><td>348 &mu;s</td><td>3 s 521 ms</td>
  </tr>
  <tr>
    <td><b>Wiki</b></td><td>238 &mu;s</td><td>241 &mu;s</td><td>17 s 514 ms</td>
  </tr>
</table>
The degree centrality calculations are very fast, however, the closeness centrality calculation is rather slow.  We should try to find some potential optimizations in the future.

Once betweenness centrality has been merged, I will add a benchmark for that as well.