@pankajgupta 2015-03-11
Filing a new issue to improve the performance of closeness centrality