@bmckown 2015-03-11
Updated.  Let me know what you think.  Thanks for filing the issue--that calculation is really rather slow.