@pankajgupta 2015-03-11
BTW, we use yourkit.com to do profiling if interested.